pyproject.toml
[tool.poetry]
name = "usv3"
version = "v3.0.0"
description = "An extensible bot framework for hack.chat"
license = "Unlicense"
authors = ["AnnikaV9 <carrot.moncher@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asyncio = "~3.4.3"
websockets = "~13.0.1"
uvloop = "~0.20.0"
loguru = "~0.7.2"
"ruamel.yaml" = "~0.18.6"
cython = { version = "^3.0.11", optional = true }
setuptools = { version = "^75.1.0", optional = true }
flake8 = { version = "^7.1.1", optional = true }

[tool.poetry.extras]
cython = ["cython", "setuptools"]
dev = ["flake8"]

[tool.poetry.scripts]
bot = "usv3.entry:main"
build_cython = "tools.tasks:build_cython"

[tool.poetry.group.cmd.dependencies]
